+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.1.5] - 2023-02-16
+
+### Added
+- Model `input_request`
+- Add `.send_stdin` method on `Kernel`
+- `Kernel` watches `stdin` channel
+
 ## [0.1.4] - 2023-02-15
 
 ### Changed